:hocho: Split tests
